feat: add various resources

+++ internal/models/account.go
@@ -7,7 +7,7 @@ import (
 type Account struct {
 	bun.BaseModel `bun:"accounts"`
 
-	AccountID int64    `bun:",pk" json:"id"`
+	AccountID int      `bun:",pk" json:"id"`
 	PenguinID string   `json:"penguinId"`
 	Weight    float64  `json:"weight"`
 	Tags      []string `json:"tags"`
