+++ internal/models/dropInfo.go
@@ -10,12 +10,12 @@ import (
 type DropInfo struct {
 	bun.BaseModel `bun:"drop_infos,alias:di"`
 
-	DropID      int64           `bun:",pk" json:"id"`
+	DropID      int             `bun:",pk" json:"id"`
 	Server      string          `json:"server"`
-	StageID     int64           `json:"stageId"`
+	StageID     int             `json:"stageId"`
 	ItemID      null.Int        `json:"itemId"`
 	DropType    string          `json:"dropType"`
-	TimeRangeID int64           `json:"timeRangeId"`
+	TimeRangeID int             `json:"timeRangeId"`
 	Bounds      json.RawMessage `json:"bounds"`
 	Accumulable bool            `json:"accumulable"`
 }
