+++ internal/models/notice.go
@@ -10,7 +10,7 @@ import (
 type Notice struct {
 	bun.BaseModel `bun:"notices"`
 
-	NoticeID   int64           `bun:",pk" json:"id"`
+	NoticeID   int             `bun:",pk" json:"id"`
 	Conditions json.RawMessage `json:"conditions"`
 	Severity   null.Int        `json:"severity"`
 	Content    string          `json:"content"`
