+++ internal/models/property.go
@@ -7,7 +7,7 @@ import (
 type Property struct {
 	bun.BaseModel `bun:"properties"`
 
-	PropertyID int64  `bun:",pk" json:"id"`
+	PropertyID int    `bun:",pk" json:"id"`
 	Key        string `json:"key"`
 	Value      string `json:"value"`
 }
