+++ internal/models/shims/stage.go
@@ -11,9 +11,9 @@ import (
 type Stage struct {
 	bun.BaseModel `bun:"stages"`
 
-	StageID      int64           `bun:",pk" json:"-"`
+	StageID      int             `bun:",pk" json:"-"`
 	ArkStageID   string          `json:"stageId"`
-	ZoneID       int64           `json:"-"`
+	ZoneID       int             `json:"-"`
 	ArkZoneID    string          `bun:"-" json:"zoneId"`
 	Code         string          `bun:"-" json:"code"`
 	CodeI18n     json.RawMessage `bun:"code" json:"code_i18n"`
