+++ internal/models/stage.go
@@ -11,11 +11,11 @@ type Stage struct {
 	bun.BaseModel `bun:"stages"`
 
 	// StageID (penguinStageId) is the numerical ID of the stage.
-	StageID int64 `bun:",pk" json:"penguinStageId"`
+	StageID int `bun:",pk" json:"penguinStageId"`
 	// ArkStageID (stageId) is the previously used, string form ID of the stage; in JSON-representation `stageId` is used as key.
 	ArkStageID string `json:"stageId"`
 	// ZoneID is the numerical ID of the zone the stage is in.
-	ZoneID int64 `json:"zoneId"`
+	ZoneID int `json:"zoneId"`
 	// Code is a map with language code as key and the code of the stage in that language as value.
 	Code json.RawMessage `json:"code"`
 	// Sanity is the sanity requirement for a full clear of the stage.
